fix(css): ensure that text buttons in selected rows remain legible
Currently in a dark theme (with dark button backgrounds), when a selected
row has a button on it, the text within that button is dark colored. With
this patch, the text should be colored based on what the theme specifies
for foreground text (usually "black" on light themes and "white" on
light themes) so that the text remains readable on dark
button backgrounds.

Fixes pop-os/gtk-theme#516